QuestCard to ts + test to testingLibrary
