+++ components/organisms-simple/QuestionCard/QuestionCard.spec.jsx
@@ -1,7 +1,10 @@
-import {mount} from 'enzyme';
+import {cleanup, render} from '@testing-library/react';
 import React from 'react';
+import {stub} from 'sinon';
 import test from 'tape';
 import QuestionCard from './QuestionCard';
+
+window.requestAnimationFrame = stub();
 
 test('QuestionCard - renders', (t) => {
   const props = {
@@ -12,15 +15,19 @@ test('QuestionCard - renders', (t) => {
     cancelChanges: () => true,
   };
 
-  const comp = mount(<QuestionCard {...props}>Content</QuestionCard>);
-
-  t.equals(comp.find('h2').first().text(), 'Question', 'Shows title');
-
-  t.equals(comp.find('p').length, 0, 'Should not show a description');
-  t.false(
-    comp.find('CardShell').first().prop('hasError'),
-    'Should not show error state',
+  const {getByText, container} = render(
+    <QuestionCard {...props}>Content</QuestionCard>,
   );
+
+  t.true(getByText('Question'), 'Shows title');
+
+  t.equals(
+    container.querySelectorAll('p').length,
+    0,
+    'Should not show a description',
+  );
+
+  cleanup();
   t.end();
 });
 
@@ -42,16 +49,21 @@ test('QuestionCard - shows description, more detail, and edit warning', (t) => {
     isLatestCard: false,
   };
 
-  const comp = mount(<QuestionCard {...props}>Content</QuestionCard>);
+  const {getByText} = render(<QuestionCard {...props}>Content</QuestionCard>);
 
-  t.equals(
-    comp.find('.uic--warning-message').length,
-    1,
+  t.true(
+    getByText(
+      'Note: Changing your answer to this question will require you to answer additional questions.',
+    ),
+
     'Should show warning text',
   );
-  t.equals(comp.find('Cabinet').length, 1, 'Should show more detail');
-  t.equals(comp.find('.uic--card-submit').first().text(), 'Save Changes');
 
+  t.true(getByText(props.moreDetails.label), 'Should show more detail');
+
+  t.true(getByText('Save Changes'), 'Should show Save Changes button');
+
+  cleanup();
   t.end();
 });
 
@@ -73,14 +85,17 @@ test('QuestionCard - collapsed', (t) => {
     cancelChanges: () => true,
   };
 
-  const comp = mount(<QuestionCard {...props}>Content</QuestionCard>);
+  const {getByText, container} = render(
+    <QuestionCard {...props}>Content</QuestionCard>,
+  );
 
-  t.equals(comp.find('h2').first().text(), 'Summary', 'Shows summary');
+  t.true(getByText('Summary'), 'Shows summary');
   t.false(
-    comp.find('CardShell').first().prop('hasError'),
+    container.querySelectorAll('.uic--card-error').length > 0,
     'Should not show error state',
   );
 
+  cleanup();
   t.end();
 });
 
@@ -103,13 +118,17 @@ test('QuestionCard - latest collapsed', (t) => {
     cancelChanges: () => true,
   };
 
-  const comp = mount(<QuestionCard {...props}>Content</QuestionCard>);
+  const {getByText, container} = render(
+    <QuestionCard {...props}>Content</QuestionCard>,
+  );
 
-  t.equals(comp.find('h2').first().text(), 'Q', 'Shows incomplete summary');
+  t.true(getByText('Q'), 'Shows incomplete summary');
+
   t.true(
-    comp.find('CardShell').first().prop('hasError'),
+    container.querySelectorAll('.uic--card-error').length > 0,
     'Should show error state',
   );
 
+  cleanup();
   t.end();
 });
